requirements.txt
flask==3.0.0
flask-cors==4.0.0
flask-sqlalchemy==3.1.1
sqlalchemy==2.0.25
pymysql==1.1.0
python-dotenv==1.0.0
PyJWT==2.8.0
bcrypt==4.1.2
requests==2.31.0
cryptography==41.0.7
